Change permutation from int to unsigned long
git-svn-id: https://ecesvn.uwaterloo.ca/courses/ece108/w18/students/mdcbonne/SatSolver@2406 bf7adb0c-0a71-4f8a-9a9b-06af373ef5e1

+++ src/main/main.c
@@ -58,7 +58,7 @@ char *classify() {
 	}
 	unsigned long numPermutations = (unsigned long)(temp - 1);
 	double truePermutations = 0;
-	int permutation;
+	unsigned long permutation;
 	for(permutation = 0; permutation <= numPermutations; permutation++){
 		// for(int j = 0; j < numVariables; j++) {
 		// 	truthValues[j] = i >> j & 1;
